MainClass
This code splits the input file into commands and processes them, updating the order book and generating responses. The results are written to the output file "output.txt". The assignment requires you to create a .jar executable with this code and use the "input.txt" file as an input file and generate the output file "output.txt".